+++ qllm/quant/__init__.py
@@ -1,4 +1,5 @@
 # from .fused_attn import QuantLlamaAttention, make_quant_attn
 # from .fused_mlp import QuantLlamaMLP, make_fused_mlp, autotune_warmup_fused
-from .quant_linear import (QuantLinear, autotune_warmup_linear)
+from .quant_linear import (QuantLinear)
+from .quant_linear_triton import autotune_warmup_linear
 # from .triton_norm import TritonLlamaRMSNorm, make_quant_norm
