+++ qllm/utils/__init__.py
@@ -1,4 +1,4 @@
 from .modelutils import DEV, find_layers, gen_conditions, torch_snr_error
 from .datautils import set_seed, get_wikitext2, get_ptb, get_c4, get_ptb_new, get_c4_new, get_loaders
 from .export import export_quant_table
-from .comm_utils import clear_memory, disable_huggingface_init
+from . import comm_utils
